Rename excel to xlsx, deleted not used components
