+++ eform-client/src/app/common/modules/eform-shared/components/eform-subheader/eform-subheader.component.html
@@ -1,6 +1,6 @@
 <div class="d-flex heading-bg">
   <div class="mr-auto p-3">
-    <h1 [ngStyle]="{ fontSize: heandingSizeRem + 'rem' }" #heading>
+    <h1 [ngStyle]="{ fontSize: heandingSizeRem + 'rem' }">
       {{ forceStaticTitle ? title : (title | translate) }}
       <small *ngIf="subtitle" style="font-size: 1rem"
         >({{ subtitle | translate }})</small
