+++ eform-client/src/app/modules/eforms/components/eforms-page/eforms-page.component.html
@@ -318,8 +318,8 @@
                   )
                 "
                 id="download-excel-btn-{{ i }}"
-                [routerLink]="['/excel-report', templateDto.id]"
-                mdbTooltip="{{ 'Download' | translate }} Excel"
+                [routerLink]="['/xlsx-report', templateDto.id]"
+                mdbTooltip="{{ 'Download' | translate }} XLSX"
               >
                 <fa-icon
                   icon="file-excel"
