--- eform-client/src/app/modules/eforms/eform-excel-report/components/eform-excel-report-container/eform-excel-report-container.component.spec.ts
@@ -1,26 +1,0 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-
-import { EformExcelReportContainerComponent } from './eform-excel-report-container.component';
-
-describe('EformExcelReportContainerComponent', () => {
-  let component: EformExcelReportContainerComponent;
-  let fixture: ComponentFixture<EformExcelReportContainerComponent>;
-
-  beforeEach(
-    waitForAsync(() => {
-      TestBed.configureTestingModule({
-        declarations: [EformExcelReportContainerComponent],
-      }).compileComponents();
-    })
-  );
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(EformExcelReportContainerComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
