--- eform-client/src/app/modules/eforms/eform-excel-report/components/eform-excel-report-header-editor/eform-excel-report-header-editor.component.html
@@ -1,119 +1,0 @@
-<div
-  [config]="{ ignoreBackdropClick: true }"
-  mdbModal
-  class="modal fade"
-  #frame="mdbModal"
->
-  <div class="modal-dialog modal-md">
-    <div class="modal-content">
-      <div class="modal-header">
-        <h3>{{ 'Edit template headers' | translate }}</h3>
-        <button class="close" data-dismiss="modal" (click)="frame.hide()">
-          <span>×</span>
-        </button>
-      </div>
-      <div class="modal-body">
-        <div class="container-fluid">
-          <div class="row">
-            <div class="col-md-12">
-              <div class="md-form">
-                <input
-                  mdbInput
-                  [(ngModel)]="reportHeadersModel.h1"
-                  required
-                  type="text"
-                  name="reportHeader1"
-                  id="reportHeader1"
-                  class="form-control"
-                />
-                <label for="reportHeader1">H1</label>
-              </div>
-            </div>
-          </div>
-          <div class="row">
-            <div class="col-md-12">
-              <div class="md-form">
-                <input
-                  mdbInput
-                  [(ngModel)]="reportHeadersModel.h2"
-                  required
-                  type="text"
-                  name="reportHeader2"
-                  id="reportHeader2"
-                  class="form-control"
-                />
-                <label for="reportHeader2">H2</label>
-              </div>
-            </div>
-          </div>
-          <div class="row">
-            <div class="col-md-12">
-              <div class="md-form">
-                <input
-                  mdbInput
-                  [(ngModel)]="reportHeadersModel.h3"
-                  required
-                  type="text"
-                  name="reportHeader3"
-                  id="reportHeader3"
-                  class="form-control"
-                />
-                <label for="reportHeader3">H3</label>
-              </div>
-            </div>
-          </div>
-          <div class="row">
-            <div class="col-md-12">
-              <div class="md-form">
-                <input
-                  mdbInput
-                  [(ngModel)]="reportHeadersModel.h4"
-                  required
-                  type="text"
-                  name="reportHeader4"
-                  id="reportHeader4"
-                  class="form-control"
-                />
-                <label for="reportHeader4">H4</label>
-              </div>
-            </div>
-          </div>
-          <div class="row">
-            <div class="col-md-12">
-              <div class="md-form">
-                <input
-                  mdbInput
-                  [(ngModel)]="reportHeadersModel.h5"
-                  required
-                  type="text"
-                  name="reportHeader5"
-                  id="reportHeader5"
-                  class="form-control"
-                />
-                <label for="reportHeader5">H5</label>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-      <div class="modal-footer">
-        <button
-          mdbBtn
-          class="btn-success btn-small-height"
-          id="saveReportHeadersBtn"
-          (click)="updateHeaders()"
-        >
-          {{ 'Update' | translate }}
-        </button>
-        <button
-          mdbBtn
-          class="btn-neutral text-black-50 btn-small-height"
-          id="cancelSaveReportHeadersBtn"
-          (click)="frame.hide()"
-        >
-          {{ 'Cancel' | translate }}
-        </button>
-      </div>
-    </div>
-  </div>
-</div>
