--- eform-client/src/app/modules/eforms/eform-excel-report/components/eform-excel-report-header/eform-excel-report-header.component.spec.ts
@@ -1,26 +1,0 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-
-import { EformExcelReportHeaderComponent } from './eform-excel-report-header.component';
-
-describe('EformExcelReportHeaderComponent', () => {
-  let component: EformExcelReportHeaderComponent;
-  let fixture: ComponentFixture<EformExcelReportHeaderComponent>;
-
-  beforeEach(
-    waitForAsync(() => {
-      TestBed.configureTestingModule({
-        declarations: [EformExcelReportHeaderComponent],
-      }).compileComponents();
-    })
-  );
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(EformExcelReportHeaderComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
