--- eform-client/src/app/modules/eforms/eform-excel-report/components/index.ts
@@ -1,3 +1,0 @@
-export * from './eform-excel-report-container/eform-excel-report-container.component';
-export * from './eform-excel-report-header/eform-excel-report-header.component';
-export * from './eform-excel-report-header-editor/eform-excel-report-header-editor.component';
