+++ eform-client/src/app/modules/eforms/eform-xlsx-report/components/eform-xlsx-report-header/eform-xlsx-report-header.component.html
@@ -20,14 +20,14 @@
         }}</label>
       </div>
     </div>
-    <div class="col-md-3 text-center pt-3">
+    <div class="col-md-1 text-center pt-3">
       <button
         mdbBtn
         class="btn-success btn-small-height"
         [disabled]="generateForm.invalid"
         (click)="onSave()"
       >
-        {{ 'Download' | translate }} Excel
+        {{ 'Download' | translate }}
       </button>
     </div>
   </div>
