+++ eform-client/src/app/modules/eforms/eform-xlsx-report/components/index.ts
@@ -1,0 +1,2 @@
+export * from './eform-xlsx-report-container/eform-xlsx-report-container.component';
+export * from './eform-xlsx-report-header/eform-xlsx-report-header.component';
