+++ eform-client/src/assets/i18n/de-DE.ts
@@ -367,4 +367,5 @@ export const deDE = {
   'Security groups': 'Beveiligingsgroepen',
   Custom: 'Op maat',
   Add: 'Toevoegen',
+  'Generate report': 'Bericht {{value}} generieren',
 };
