+++ eform-client/src/assets/i18n/en-US.ts
@@ -359,4 +359,5 @@ export const enUS = {
   'Security groups': 'Security groups',
   Custom: 'Custom',
   Add: 'Add',
+  'Generate report': 'Generate {{value}} report',
 };
